move AnimalList and AnimalCard to AnimalList directory, add host profile and animal list to host profile view
