+++ src/Components/HostProfileView/HostProfile.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+export const HostProfile = () => (
+  <div
+    style={{
+      display: "inline-block",
+      backgroundColor: "BurlyWood",
+      height: 200,
+      width: "40vw"
+    }}
+  >
+    Host Profile
+  </div>
+);
